+++ ContentPatcher/Framework/PatchManager.cs
@@ -148,12 +148,13 @@ namespace ContentPatcher.Framework
         /// <param name="date">The current in-game date (if applicable).</param>
         /// <param name="weather">The current in-game weather (if applicable).</param>
         /// <param name="spouse">The current player's internal spouse name (if applicable).</param>
-        public void UpdateContext(IContentHelper contentHelper, LocalizedContentManager.LanguageCode language, SDate date, Weather? weather, string spouse)
+        /// <param name="dayEvent">The day event (e.g. wedding or festival) occurring today (if applicable).</param>
+        public void UpdateContext(IContentHelper contentHelper, LocalizedContentManager.LanguageCode language, SDate date, Weather? weather, string dayEvent, string spouse)
         {
             this.VerboseLog("Propagating context...");
 
             // update context
-            this.ConditionContext.Set(language, date, weather, spouse);
+            this.ConditionContext.Set(language: language, date: date, weather: weather, dayEvent: dayEvent, spouse: spouse);
             IDictionary<ConditionKey, string> tokenisableConditions = this.ConditionContext.GetSingleValueConditions();
 
             // update patches